+++ actix-utils/src/timeout.rs
@@ -85,8 +85,8 @@ where
     type Request = S::Request;
     type Response = S::Response;
     type Error = TimeoutError<S::Error>;
-    type InitError = E;
     type Transform = TimeoutService<S>;
+    type InitError = E;
     type Future = Ready<Result<Self::Transform, Self::InitError>>;
 
     fn new_transform(&self, service: S) -> Self::Future {
@@ -146,6 +146,7 @@ where
 pub struct TimeoutServiceResponse<T: Service> {
     #[pin]
     fut: T::Future,
+    #[pin]
     sleep: Sleep,
 }
 
@@ -156,7 +157,7 @@ where
     type Output = Result<T::Response, TimeoutError<T::Error>>;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let mut this = self.project();
+        let this = self.project();
 
         // First, try polling the future
         match this.fut.poll(cx) {
@@ -166,7 +167,7 @@ where
         }
 
         // Now check the sleep
-        match Pin::new(&mut this.sleep).poll(cx) {
+        match this.sleep.poll(cx) {
             Poll::Pending => Poll::Pending,
             Poll::Ready(_) => Poll::Ready(Err(TimeoutError::Timeout)),
         }