+++ actix-codec/CHANGES.md
@@ -2,8 +2,9 @@
 
 ## Unreleased - 2020-xx-xx
 * Upgrade `pin-project` to `1.0`.
-* Update `tokio` dependency to 0.3.1.
-* Update `tokio-util` dependency to 0.5.1.
+* Upgrade `tokio` dependency to `1`.
+* Upgrade `tokio-util` dependency to `0.6`.
+* Upgrade `bytes` dependency to `1`.
 
 ## 0.3.0 - 2020-08-23
 * No changes from beta 2.