+++ actix-codec/CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
-* Upgrade `pin-project` to `1.0`.
+* Replace `pin-project` with `pin-project-lite`.
+* Upgrade `tokio` dependency to `1`.
+* Upgrade `tokio-util` dependency to `0.6`.
+* Upgrade `bytes` dependency to `1`.
 
 ## 0.3.0 - 2020-08-23
 * No changes from beta 2.