+++ actix-codec/CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
-* `Framed::poll_ready` flushes when the buffer is full [#431]
+* `Framed::poll_ready` flushes when the buffer is full [#409]
 
 ## 0.4.0 - 2021-04-20
 * No significant changes since v0.4.0-beta.1.